Remove function dealing with expanded release version
Conflating vcs versions with Erlang release versions is not a good
idea. This may have gone unnoticed for some time since the vcs version
detection has been broken and only the Erlang release version has been
used. Remove the functions dealing with expanding the release version
and use the version as specified in the reltool configuration.