Improve the check to determine if an application requires building
Add an application-level check for the need to build an application so
that unchanged applications can be effectively skipped. This improves
the efficiency of the build process where many or all of the
applications involved have not been modified. All erlang source and
include files are gathered and the modified times compared to the
modified time of the ebin directory. If any source or include files
have changed since the ebin directory was modified then a normal
compile is done. Since beam files are only created or removed from the
ebin directory the modified time of the directory should always be
representative of the last build and there should not be a practical
case where an application would be skipped when source or include
files have changed.